==> 01_DomainBounds2json.ipynb <==
# create a JSON with bounding parameters for each modeling domain
# get ncols, nrows, ll corner coordinates, and cell size from domain ascii files. 

# {'latmax' : maxy,'latmin' : miny,'lonmax': maxx,'lonmin': minx,}
BEAU = {
    "name": 'BEAU',
    "Bbox": {'latmax' : 71.34,'latmin' : 67.14,'lonmax': -136.48,'lonmin': -166.51,},
    "st": "1988-10-01", 
    "ed": "2019-09-30",
    "stn_proj": 'epsg:4326',
    "mod_proj": 'epsg:3338',
    "ncols": '1298',
    "nrows": '496',
    "xll": '-543000',
    "yll": '1960000', 
    "cellsize":'1000' 
}

BRIS = {
    "name": 'BRIS',
    "Bbox": {'latmax' : 64.32,'latmin' : 53.74,'lonmax': -151.31,'lonmin': -165.71,},
    "st": "1988-10-01", 
    "ed": "2019-09-30",
    "stn_proj": 'epsg:4326',
    "mod_proj": 'epsg:3338',
    "ncols": '949',
    "nrows": '1228',
    "xll": '-771000',
    "yll": '417000', 
    "cellsize":'1000' 
}

CHUK = {
    "name": 'CHUK',
    "Bbox": {'latmax' : 68.82,'latmin' : 62.23,'lonmax': -153.64,'lonmin': -168.45,},
    "st": "1988-10-01", 
    "ed": "2019-09-30",
    "stn_proj": 'epsg:4326',
    "mod_proj": 'epsg:3338',
    "ncols": '763',
    "nrows": '798',
    "xll": '-744000',
    "yll": '1361000', 
    "cellsize":'1000' 
}

JUNE = {
    "name": 'JUNE',
    "Bbox": {'latmax' : 58.677,'latmin' : 57.700,'lonmax': -133.741,'lonmin': -136.640,},
    "st": "1988-10-01", 
    "ed": "2019-09-30",
    "stn_proj": 'epsg:4326',
    "mod_proj": 'epsg:26708',
    "ncols": '692',
    "nrows": '438',
    "xll": '402250',
    "yll": '6395750', 
    "cellsize":'250'
}


KLAW = {
    "name": 'KLAW',
    "Bbox": {'latmax' : 55.975,'latmin' : 55.015,'lonmax': -131.755,'lonmin': -134.450,},
    "st": "1988-10-01", 
    "ed": "2019-09-30",
    "stn_proj": 'epsg:4326',
    "mod_proj": 'epsg:26708',
    "ncols": '694',
    "nrows": '447',
    "xll": '534250',
    "yll": '6096250', 
    "cellsize":'250' 
}

SITK = {
    "name": 'SITK',
    "Bbox": {'latmax' : 57.587,'latmin' : 56.590,'lonmax': -134.275,'lonmin': -137.085,},
    "st": "1988-10-01", 
    "ed": "2019-09-30",
    "stn_proj": 'epsg:4326',
    "mod_proj": 'epsg:26708',
    "ncols": '691',
    "nrows": '451',
    "xll": '372000',
    "yll": '6271750', 
    "cellsize":'250'
}

TAKU = {
    "name": 'TAKU',
    "Bbox": {'latmax' : 59.60,'latmin' : 57.75,'lonmax': -130.70,'lonmin': -134.90,},
    "st": "1988-10-01", 
    "ed": "2019-09-30",
    "stn_proj": 'epsg:4326',
    "mod_proj": 'epsg:26708',
    "ncols": '1002',
    "nrows": '857',
    "xll": '505500',
    "yll": '6400500', 
    "cellsize":'250' 
}

YUKO = {
    "name": 'YUKO',
    "Bbox": {'latmax' : 68.98,'latmin' : 58.25,'lonmax': -129.0,'lonmin': -164.67,},
    "st": "1988-10-01", 
    "ed": "2019-09-30",
    "stn_proj": 'epsg:4326',
    "mod_proj": 'epsg:3338',
    "ncols": '2052',
    "nrows": '1336',
    "xll": '-622000',
    "yll": '966000', 
    "cellsize":'1000' 
}


NPRB_domains = {
    "BEAU": BEAU,
    "BRIS": BRIS,
    "CHUK": CHUK,
    "JUNE": JUNE,
    "KLAW":KLAW,
    "SITK":SITK,
    "TAKU":TAKU,
    "YUKO":YUKO,
}


NPRB_domains.keys()

import json
json = json.dumps(NPRB_domains)
f = open('/nfs/attic/dfh/2020_NPRB/notebooks/02_preprocess_python/NPRB_domains.json',"w")
f.write(json)
f.close()

# #save baseline .par file
# import json

# json = json.dumps(CSO_domains)
# f = open('/nfs/attic/dfh/Aragon2/CSOdata/CSO_domains.json',"w")
# f.write(json)
# f.close()
