==> 07_Baseline_par2json.ipynb <==
####### USER #########
#select domain
domain = 'BEAU'

#path to parfile 
parFile = '/nfs/attic/dfh/2020_NPRB/domain_'+domain+'/test/snowmodel.par'

# outfile path
outpath = '/nfs/attic/dfh/2020_NPRB/data/json/par_base.json'

#create dictionay with values from text file - may want to use for calibration
def getparams(file_name):
    myvars = {}
    lines = open(file_name, 'r').readlines()
    for i in range(len(lines)):
        line=lines[i]
        if i == 14 or i == 17 or i == 18 or i == 19 or i == 93 or i == 95 \
        or i == 97 or i == 100 or i == 102 or i == 104 or i == 107 or i == 108 \
        or i ==147 or i == 148 or i == 149:
            var = line.strip()
            if i == 14:
                name = 'met_input_fname'
            elif i == 17:
                name = 'topoveg_grads_fname'
            elif i == 18:
                name = 'topo_ascii_fname'
            elif i == 19:
                name = 'veg_ascii_fname'
            elif i == 93:
                name = 'micromet_output_fname'
            elif i == 95:
                name = 'snowtran_output_fname'
            elif i == 97:
                name = 'tabler_sfc_path_name'
            elif i == 100:
                name = 'enbal_output_fname'
            elif i == 102:
                name = 'snowpack_output_fname'
            elif i == 104:
                name = 'multilayer_output_fname'
            elif i == 107:
                name = 'output_path_wo_assim'
            elif i == 108:
                name = 'output_path_wi_assim'
            elif i == 147:
                name = 'lat_file_path'
            elif i == 148:
                name = 'lon_file_path'
            elif i == 149:
                name = 'snowmodel_line_file'
            myvars[name] = [var,name,i]
        else:
            var, name = line.partition("!")[::2]
            #create a dictionary with the variable name as the key for the 
            #[model value, parFile text, linenumber]
            myvars[name.split()[0]] = [var.strip(),name,i]
    return myvars
base = getparams(parFile)
#print(base)
print(base['nx'])

for key in base:
    print(key)
    print(base[key])

#save baseline .par file
import json

json = json.dumps(base)
f = open(outpath,"w")
f.write(json)
f.close()
