==> 11_Build_SnowModel_line_file.ipynb <==
import numpy as np
import geopandas as gpd
import pandas as pd
import codecs
import requests
import rasterio as rio 
from rasterstats import zonal_stats, point_query

#########################################################################
############################ USER INPUTS ################################
#########################################################################

# DOMAIN
# choose the modeling domain
domain = 'BEAU'

# PATHS
# path to domain data folder 
dataPath = '/nfs/attic/dfh/2020_NPRB/data/SMinputs/'+domain+'/'
#########################################################################

def build_snowmodel_line(domain,stn_gdf,dataPath):

    # path to DEM ascii
    DEMpath = dataPath+domain+'_dem.asc'

    #path to VEG .asc
    VEGpath = dataPath+domain+'_veg.asc'

    #path to lat .asc
    LATpath = dataPath+domain+'_grid_lat.asc'

    #path to lon .asc
    LONpath = dataPath+domain+'_grid_lon.asc'

    # VEG outfile path
    outVEGpath = dataPath+domain+'_veg_line.asc'

    # DEM outfile path
    outDEMpath = dataPath+domain+'_dem_line.asc'

    # Line outfile path
    outFpath = dataPath+'/snowmodel_line_pts.dat'

    #lon outfile path
    outLONpath = dataPath+domain+'_grid_lon_line.asc'

    #lat outfile path
    outLATpath = dataPath+domain+'_grid_lat_line.asc'

    #path to CSO domain
    domains_resp = requests.get("https://raw.githubusercontent.com/NPRB/02_preprocess_python/main/NPRB_domains.json")
    domains = domains_resp.json()
    # CSO projection
    mod_proj = domains[domain]['mod_proj']

    #get metadata from .asc 
    def getmeta(file_name):
        myvars = {}
        lines = open(file_name, 'r').readlines()
        for i in range(5):
            line=lines[i]
            name, var = line.partition("\t")[::2]
            myvars[name.strip()] = int(var)
        return myvars

    myvars = getmeta(DEMpath)

    #Build northing easting array 
    #Northing
    # locate center of lower left cell
    st = myvars['yllcorner']+(myvars['cellsize']/2)
    #build northing array
    north = np.arange(st,st+myvars['nrows']*myvars['cellsize'],myvars['cellsize'])

    #Easting
    # locate center of lower left cell
    st = myvars['xllcorner']+(myvars['cellsize']/2)
    #build easting array
    east = np.arange(st,st+myvars['ncols']*myvars['cellsize'],myvars['cellsize'])

    #fortran indexing starts at 1
    #staion 1 to N of N stations
    count = np.zeros(stn_gdf.shape[0])
    #easting of pixel corresponding to station 
    stn_est = np.zeros(stn_gdf.shape[0])
    #northing of pixel corresponding to station
    stn_nor = np.zeros(stn_gdf.shape[0])
    #index of pixel corresponding to station 
    est_idx = np.zeros(stn_gdf.shape[0])
    #index of pixel corresponding to station 
    nor_idx = np.zeros(stn_gdf.shape[0])

    for z in range(stn_gdf.shape[0]):
        count[z] = z + 1
        lons = abs(stn_gdf.easting[z]-east)
        loIDX = [i for i, value in enumerate(lons) if value == np.min(abs(stn_gdf.easting[z]-east))] 
        stn_est[z] = east[loIDX[0]]
        est_idx[z] = loIDX[0] + 1
        lats = abs(stn_gdf.northing[z]-north)
        laIDX = [i for i, value in enumerate(lats) if value == np.min(abs(stn_gdf.northing[z]-north))]
        stn_nor[z] = north[laIDX[0]]
        nor_idx[z] = laIDX[0] + 1

    #Print out .dat file 
    f= open(outFpath,"w+")
    for z in range(count.shape[0]):
        f.write('{:08.0f}\t'.format(count[z])+'{:10.0f}\t'.format(est_idx[z])+'{:10.0f}\t'.format(nor_idx[z])+
                '{:10.0f}\t'.format(stn_est[z])+'{:10.0f}\t\n'.format(stn_nor[z]))
    f.close() 

    ## Extract topo, veg, lat, and lon files for snowmodel_line
    new = stn_gdf.to_crs(mod_proj)

    with rio.open(DEMpath) as src:
        rows, cols = rio.transform.rowcol(src.transform, new.geometry.centroid.x, new.geometry.centroid.y)

    # function to extract .asc files at SNOTEL locations
    def field2line(filepath,outpath):
        data_line=[]
        for i in range(stn_gdf.shape[0]):
            info = str(point_query([new.geometry[i]], filepath)[0])
            data_line.append(info)
        lines = open(filepath, 'r').readlines()
        head = 'ncols\t1\nnrows\t'+str(stn_gdf.shape[0])+'\n'+lines[2]+lines[3]+lines[4]+lines[5]
        data = ''
        for s in data_line:
            data += s +'\n'
        f= open(outpath,"w+")
        f.write(head+data)
        f.close()    

    #extract DEM
    field2line(DEMpath,outDEMpath)
    #extract VEG
    field2line(VEGpath,outVEGpath)
    #extract LAT
    field2line(LATpath,outLATpath)
    #extract LON
    field2line(LONpath,outLONpath)

domain_list = ['BEAU']

for domain in domain_list:
    print(domain)
    
    dataPath = '/nfs/attic/dfh/2020_NPRB/data/SMinputs/'+domain+'/'
    #path to SNOTEL gdf
    spath = '/nfs/attic/dfh/2020_NPRB/data/snotel/'+domain+'/Station_'+domain+'.geojson'
    
    #path to snow course gdf
    scpath = '/nfs/attic/dfh/2020_NPRB/data/snow_course/'+domain+'/Snow_course_'+domain+'.geojson'
    
    if (domain == 'KLAW')|(domain == 'SITK'):
        # these domains have no data
        continue
    if domain == 'CHUK':
        # this domain only has snow course data
        stn_gdf = gpd.read_file(scpath)
        #build input file
        build_snowmodel_line(domain,stn_gdf,dataPath)
    else:
        # station data
        stn_gdf = gpd.read_file(spath)
#         # station data
#         sc_gdf = gpd.read_file(scpath)
#         sc_gdf=sc_gdf.rename(columns={"lat": "latitude", "lon": "longitude"})
#         #combine geodataframes
#         stn_gdf = pd.concat([st_gdf,sc_gdf]).reset_index(drop=True)
        #build input file
        build_snowmodel_line(domain,stn_gdf,dataPath)

domain_list = ['BEAU','BRIS','CHUK','JUNE','KLAW','SITK','TAKU','YUKO']

for domain in domain_list:
    print(domain)
    
    dataPath = '/nfs/attic/dfh/2020_NPRB/data/SMinputs/'+domain+'/'
    #path to SNOTEL gdf
    spath = '/nfs/attic/dfh/2020_NPRB/data/snotel/'+domain+'/Station_'+domain+'.geojson'
    
    #path to snow course gdf
    scpath = '/nfs/attic/dfh/2020_NPRB/data/snow_course/'+domain+'/Snow_course_'+domain+'.geojson'
    
    if (domain == 'KLAW')|(domain == 'SITK'):
        # these domains have no data
        continue
    if domain == 'CHUK':
        # this domain only has snow course data
        stn_gdf = gpd.read_file(scpath)
        #build input file
        build_snowmodel_line(domain,stn_gdf,dataPath)
    else:
        # station data
        stn_gdf = gpd.read_file(spath)
#         # station data
#         sc_gdf = gpd.read_file(scpath)
#         sc_gdf=sc_gdf.rename(columns={"lat": "latitude", "lon": "longitude"})
#         #combine geodataframes
#         stn_gdf = pd.concat([st_gdf,sc_gdf]).reset_index(drop=True)
        #build input file
        build_snowmodel_line(domain,stn_gdf,dataPath)
